refactor:入力フォームをクラスに分割
